fix: round every floating-point product so arm64 matches amd64
On arm64 (and ppc64le, s390x, riscv64, loong64) the Go compiler fuses a
floating-point product that feeds an addition or subtraction into a fused
multiply-add, also across statements, through x *= y, through inlined
calls, and through divisions by powers of two. The fused result rounds
differently from amd64 and from libopus, so the float32 ports (FFT, MDCT,
CELT and SILK analysis) produced different packets on arm64: the libopus
float-bit FFT/MDCT tests and the encoder output digests failed in the arm64
CI jobs.

Every product that reached an addition is now rounded by an explicit
conversion, which the Go spec guarantees and which compiles to nothing on
amd64; an arm64 build of the module contains no fused multiply-add. The
amd64 output is unchanged (go test and go test -tags opusref pass as
before). CI now fails when an arm64 build contains one, and
docs/DEVELOPER.md states the rule.

+++ bandwidth_detect.go
@@ -74,7 +74,7 @@ func detectSignalBandwidthAndSparsity(pcm []float64, channels, sampleRate, prev 
 	var emax, sum, sumSquares float64
 	for k := 1; k < len(power); k++ {
 		sum += power[k]
-		sumSquares += power[k] * power[k]
+		sumSquares += float64(power[k] * power[k])
 		if power[k] > emax {
 			emax = power[k]
 		}
@@ -112,7 +112,7 @@ func isSpectrallySparse(pcm []float64, channels int) bool {
 	var sum, sumSquares float64
 	for k := 1; k < len(power); k++ {
 		sum += power[k]
-		sumSquares += power[k] * power[k]
+		sumSquares += float64(power[k] * power[k])
 	}
 	if sum == 0 || sumSquares == 0 {
 		return false
@@ -149,7 +149,7 @@ func signalSpectrumPower(pcm []float64, channels int) ([]float64, int, bool) {
 	// Hann window so a strong low tone does not leak enough into high bins to look
 	// like real high-frequency content, then zero-pad to a power-of-two FFT size.
 	for i := 0; i < n; i++ {
-		w := 0.5 - 0.5*math.Cos(2*math.Pi*float64(i)/float64(n-1))
+		w := 0.5 - float64(0.5*math.Cos(2*math.Pi*float64(i)/float64(n-1)))
 		mono[i] *= w
 	}
 	nfft := dsp.NextPowerOf2(n)
@@ -163,7 +163,7 @@ func signalSpectrumPower(pcm []float64, channels int) ([]float64, int, bool) {
 
 	power := make([]float64, len(spec))
 	for k := 1; k < len(spec); k++ {
-		p := spec[k].Real*spec[k].Real + spec[k].Imag*spec[k].Imag
+		p := float64(spec[k].Real*spec[k].Real) + float64(spec[k].Imag*spec[k].Imag)
 		power[k] = p
 	}
 	return power, nfft, true
